- Use long as modelica_integer type
git-svn-id: https://openmodelica.org/svn/OpenModelica/trunk@10396 f25d12d1-65f4-0310-ae8a-bbce733d8d8e

+++ SimulationRuntime/c/openmodelica.h
@@ -119,7 +119,7 @@ typedef modelica_string_const modelica_string;
 
 /* BEFORE: #include "memory_pool.h" */
 typedef double      m_real;
-typedef int        m_integer;
+typedef long        m_integer;
 typedef const char* m_string;
 typedef signed char m_boolean;
 typedef m_integer   _index_t;
